Optimize includes a bit
