+++ src/LoginPage.cpp
@@ -15,6 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <QPainter>
 #include <QStyleOption>
 
 #include <mtx/identifiers.hpp>
