+++ src/TopRoomBar.h
@@ -17,23 +17,21 @@
 
 #pragma once
 
-#include <QAction>
-#include <QIcon>
-#include <QImage>
-#include <QLabel>
-#include <QPaintEvent>
-#include <QPainter>
-#include <QPen>
-#include <QPoint>
-#include <QStyle>
-#include <QStyleOption>
-#include <QVBoxLayout>
+#include <QColor>
+#include <QStringList>
+#include <QWidget>
 
 class Avatar;
 class FlatButton;
 class Menu;
 class TextLabel;
 class OverlayModal;
+
+class QPainter;
+class QLabel;
+class QIcon;
+class QHBoxLayout;
+class QVBoxLayout;
 
 class TopRoomBar : public QWidget
 {
@@ -67,19 +65,8 @@ signals:
         void mentionsClicked(const QPoint &pos);
 
 protected:
-        void mousePressEvent(QMouseEvent *) override
-        {
-                if (roomSettings_ != nullptr)
-                        roomSettings_->trigger();
-        }
-
-        void paintEvent(QPaintEvent *) override
-        {
-                QStyleOption opt;
-                opt.init(this);
-                QPainter p(this);
-                style()->drawPrimitive(QStyle::PE_Widget, &opt, &p, this);
-        }
+        void mousePressEvent(QMouseEvent *) override;
+        void paintEvent(QPaintEvent *) override;
 
 private:
         QHBoxLayout *topLayout_  = nullptr;
