+++ src/emoji/Panel.cpp
@@ -15,6 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <QPainter>
 #include <QPushButton>
 #include <QScrollBar>
 #include <QVBoxLayout>
